Thêm api cho employer và job

+++ src/main/java/vn/nhantd/mycareer/model/user/Salary.java
@@ -2,8 +2,8 @@ package vn.nhantd.mycareer.model.user;
 
 public class Salary {
     private String currency_unit;
-    private Double from_salary;
-    private Double to_salary;
+    private Long from_salary;
+    private Long to_salary;
 
     // Standard getters & setters
     public String getCurrency_unit() {
@@ -14,19 +14,19 @@ public class Salary {
         this.currency_unit = currency_unit;
     }
 
-    public Double getFrom_salary() {
+    public Long getFrom_salary() {
         return from_salary;
     }
 
-    public void setFrom_salary(Double from_salary) {
+    public void setFrom_salary(Long from_salary) {
         this.from_salary = from_salary;
     }
 
-    public Double getTo_salary() {
+    public Long getTo_salary() {
         return to_salary;
     }
 
-    public void setTo_salary(Double to_salary) {
+    public void setTo_salary(Long to_salary) {
         this.to_salary = to_salary;
     }
